requirements.txt
asgiref==3.5.1
black==22.3.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
dj-database-url==0.5.0
Django==3.1.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.5.2
pytz==2022.1
requests==2.27.1
semantic-version==2.10.0
setuptools-rust==1.3.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.1.0
